ESM_2-Modern Implementation/requirements.txt
torch
transformers
scikit-learn
pandas
numpy
tqdm
biopython
matplotlib
seaborn
fair-esm
biotite
omegaconf
